New navigation, and converting the homepage to a gatsby page, by importing the HTML, CSS, and JS properly
Now deployment will work properly because all assets have hash names, cookie consent is
managed by the gatsby site so it will work on all pages properly, the navbar is no longer
duplicated between index.html and /components/Nav/, the static directory is cleaned up,
and the gatsby develop command works properly now. Remaining issue is to scope the guide
JS so it doesn't throw errors when it's running on the wrong page (because it is still
all global—that will be addressed when the guide page becomes an automated React page).